test(LevelTest): rename test class and refactor method
