+++ src/main/java/com/dinosaur/dinosaurexploder/utils/LevelManager.java
@@ -7,7 +7,7 @@ package com.dinosaur.dinosaurexploder.utils;
  */
 public class LevelManager {
     private int currentLevel = 1;
-    private int enemiesToDefeat = 5; 
+    private int enemiesToDefeat = 5;
     private int defeatedEnemies = 0;
     private double enemySpawnRate = 0.75;
     private double enemySpeed = 1.5;
@@ -19,6 +19,7 @@ public class LevelManager {
     public double getEnemySpawnRate() {
         return enemySpawnRate;
     }
+
     public double getEnemySpeed() {
         return enemySpeed;
     }
@@ -38,5 +39,9 @@ public class LevelManager {
 
         enemySpawnRate = Math.max(0.3, enemySpawnRate * 0.9);
         enemySpeed += 0.2;
+    }
+
+    public int getEnemiesToDefeat() {
+        return enemiesToDefeat;
     }
 }
